local/requirements.txt
ibapi==10.19.2
requests>=2.31.0
python-dotenv>=1.0.0
